created new project
CoreModel-based
