+++ Cartfile
@@ -1,1 +1,3 @@
-github "ColemanCDA/RoutingHTTPServer" "develop"
+github "PureSwift/SwiftFoundation" "develop"
+github "PureSwift/CoreModel" "develop"
+github "PureSwift/libwebsockets" "master"
