+++ NetworkObjects/NetworkObjectsTests/NetworkObjectsTests.swift
@@ -2,14 +2,12 @@
 //  NetworkObjectsTests.swift
 //  NetworkObjectsTests
 //
-//  Created by Alsey Coleman Miller on 9/12/14.
-//  Copyright (c) 2014 ColemanCDA. All rights reserved.
+//  Created by Alsey Coleman Miller on 9/1/15.
+//  Copyright © 2015 ColemanCDA. All rights reserved.
 //
 
-import Foundation
 import XCTest
-import CoreData
-import NetworkObjects
+@testable import NetworkObjects
 
 class NetworkObjectsTests: XCTestCase {
     
@@ -23,31 +21,16 @@ class NetworkObjectsTests: XCTestCase {
         super.tearDown()
     }
     
-    func testFrameworkVersion() {
-                
-        let shortFrameworkVersion = NetworkObjectsFrameworkBundle.infoDictionary!["CFBundleShortVersionString"] as! String
-        
-        let frameworkVersion = NetworkObjectsFrameworkBundle.infoDictionary![kCFBundleVersionKey as String]! as! String
-        
-        if frameworkVersion == "TRAVISCI" {
-            
-            print("Skipping framework version assertion, running from Travis CI")
-            
-            return
-        }
-        
-        print("Testing NetworkObjects \(shortFrameworkVersion) Build \(frameworkVersion)")
-        
-        XCTAssert(shortFrameworkVersion != "1", "Short framework version (\(shortFrameworkVersion)) should not equal 1")
-        
-        XCTAssert(UInt(frameworkVersion) > 1, "Framework version (\(frameworkVersion)) should be greater than 1")
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
     }
     
-    func testStoreInit() {
-        // This is an example of a functional test case.
-        
-        let store = Store(managedObjectModel: NSManagedObjectModel(), managedObjectContextConcurrencyType: NSManagedObjectContextConcurrencyType.MainQueueConcurrencyType, serverURL: NSURL(string: "http://localhost:8080")!, prettyPrintJSON: true, resourceIDAttributeName: "id", dateCachedAttributeName: "dateCached", searchPath: "search")
-        
-        XCTAssert(store.dynamicType === Store.self, "Store should have been initialized")
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measureBlock {
+            // Put the code you want to measure the time of here.
+        }
     }
+    
 }
