+++ Old/NetworkObjects/NetworkObjects.xcodeproj/project.xcworkspace/xcuserdata/Coleman.xcuserdatad/WorkspaceSettings.xcsettings
