+++ Old/NetworkObjects/NetworkObjects.xcodeproj/xcuserdata/alseycolemanmiller.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>NetworkObjects OS X.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>2</integer>
+		</dict>
+		<key>NetworkObjects iOS.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>6E17985F1B323A61005D2844</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>6E1798731B323A91005D2844</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>6EAD584519C3B58300DC080A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>6EAD585019C3B58300DC080A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
