+++ Old/NetworkObjects/NetworkObjects/ConnectionType.swift
