+++ Old/NetworkObjects/NetworkObjects/CoreDataAttributeJSONCompatibility.swift
