+++ Old/NetworkObjects/NetworkObjects/CoreDataFetchRequestJSONCompatibility.swift
