+++ Old/NetworkObjects/NetworkObjects/CoreDataPredicateJSONCompatibility.swift
