+++ Old/NetworkObjects/NetworkObjects/Define.swift
