+++ Old/NetworkObjects/NetworkObjects/ErrorValue.swift
