+++ Old/NetworkObjects/NetworkObjects/RequestType.swift
