+++ Old/NetworkObjects/NetworkObjects/StoreError.swift
