+++ Old/NetworkObjects/NetworkObjectsTests/CoreDataAttributeJSONCompatibilityTests.swift
