+++ Old/NetworkObjects/NetworkObjectsTests/TestModel.xcdatamodeld/TestModel.xcdatamodel/contents
