+++ Source/RequestType.swift
@@ -1,0 +1,29 @@
+//
+//  RequestType.swift
+//  NetworkObjects
+//
+//  Created by Alsey Coleman Miller on 9/1/15.
+//  Copyright © 2015 ColemanCDA. All rights reserved.
+//
+
+/** Request Type */
+public enum RequestType {
+    
+    /** GET request */
+    case Get
+    
+    /** PUT (edit) request */
+    case Edit
+    
+    /** DELETE request */
+    case Delete
+    
+    /** POST (create new) request */
+    case Create
+    
+    /** Search request */
+    case Search
+    
+    /** Function request */
+    case Function
+}
