+++ Source/Server.swift
@@ -1,0 +1,15 @@
+//
+//  Server.swift
+//  NetworkObjects
+//
+//  Created by Alsey Coleman Miller on 9/1/15.
+//  Copyright © 2015 ColemanCDA. All rights reserved.
+//
+
+import SwiftFoundation
+import CoreModel
+
+public class Server {
+    
+    
+}
